refactor(core): Split transform implementation in smaller modules
